Master ProcessdataMapper Mapper

+++ src/main/java/de/vrees/heatpump/slaves/beckhoff/EL3064.java
@@ -53,7 +53,7 @@ public class EL3064 extends Slave {
     }
 
 
-    public int getPressureDiffenceEvaporator() {
+    public float getPressureDiffenceEvaporator() {
         return pressureDiffenceEvaporator.value.get();
     }
